show correct progress on textviews

+++ app/src/main/java/org/nutritionfacts/dailydozen/Servings.java
@@ -1,5 +1,5 @@
 package org.nutritionfacts.dailydozen;
 
 public interface Servings {
-    int getServings();
+    float getServings();
 }
